requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gevent==24.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.1.1
PyAudio==0.2.14
pydub==0.25.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
setuptools==75.1.0
termcolor==2.4.0
urllib3==2.2.3
webrtcvad==2.0.10
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
zope.event==5.0
zope.interface==7.0.3
